+++ includes/functions-display.php
@@ -558,6 +558,50 @@ function ul_get_opening_hours( $args ) {
 }
 
 /**
+ * Get location's Google Analytics code
+ * Only gets code on location content
+ *
+ * @since  1.2.4
+ *
+ * @return mixed
+ */
+function ul_get_location_ga_code() {
+	$code = '';
+	// Handles ul_is_location_content() check
+	$location_id = ul_get_location_id();
+	if ( $location_id ) {
+		$code = ul_get_ga_code( $location_id );
+	}
+	return $code;
+}
+
+/**
+ * Get location's Google Analytics code
+ * from a location parent page ID
+ *
+ * @since  1.2.4
+ *
+ * @return mixed
+ */
+function ul_get_ga_code( $location_parent_id ) {
+	// Get the location
+	$tracking_id = ul_get_field( $location_parent_id, 'ga_tracking_id' );
+	if ( ! $tracking_id ) {
+		return;
+	}
+
+	return "<script>
+			(function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
+			(i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
+			m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
+			})(window,document,'script','https://www.google-analytics.com/analytics.js','ga');
+			ga('create', '{$tracking_id}', 'auto');
+			ga('send', 'pageview');
+		</script>";
+
+}
+
+/**
  * TODO
  */
 function ul_get_map( $location_parent_id ) {